Add a basic test sequence to check the "happy path".
TODO: ensure that the model can actually test an interesting range of
behaviour.
TODO: store email records in a global variable, instead of the SQLite
database, when carrying out tests. This means that after sending
requests to routes that send emails in the background, we do not need
to enforce a delay (e.g. sleeping on the thread) before reading from
the emails table (as the background thread may not have been
scheduled before the test thread reads the value from the table).

Interesting question: how long a delay is necessary (I imagine this
depends on the specific thread scheduler - e.g. setting a delay of
1ms may be enough to ensure the background thread is scheduled before
our thread on some systems, but on others the background thread
might not complete)?